# contract/openapi.json
{
  "components": {
    "schemas": {
...
      "ConversationEnvelope": {
        "additionalProperties": false,
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/ConversationOut"
          }
        },
        "required": [
          "conversation"
        ],
        "title": "ConversationEnvelope",
        "type": "object"
      },
...
      "ConversationOut": {
        "additionalProperties": false,
        "properties": {
          "agentId": {
            "title": "Agentid",
            "type": "string"
          },
          "collectionId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collectionid"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lastRunId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastrunid"
          },
          "ownerUserId": {
            "format": "uuid",
            "title": "Owneruserid",
            "type": "string"
          },
          "taskId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taskid"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ownerUserId",
          "agentId",
          "title",
          "lastRunId",
          "taskId",
          "collectionId",
          "createdAt",
          "updatedAt"
        ],
        "title": "ConversationOut",
        "type": "object"
      },
...
      "ConversationTaskIn": {
        "additionalProperties": false,
        "description": "把这段对话记在某张需求单下，或者摘掉（给 ``null``）。理由同上，单独一个端点。",
        "properties": {
          "taskId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taskid"
          }
        },
        "required": [
          "taskId"
        ],
        "title": "ConversationTaskIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TaskBrief": {
        "additionalProperties": false,
        "description": "一份需求单上的创作输入。\n\n每一项都可以先空着——需求方通常是分几次填完的，草稿阶段不催。发布时才要求它至少\n说清楚要做什么（见 ``service.py`` 的发布关卡）。",
        "properties": {
          "audience": {
            "default": "",
            "maxLength": 200,
            "title": "Audience",
            "type": "string"
          },
          "color": {
            "default": "",
            "maxLength": 200,
            "title": "Color",
            "type": "string"
          },
          "contentType": {
            "default": "",
            "maxLength": 200,
            "title": "Contenttype",
            "type": "string"
          },
          "department": {
            "default": "",
            "maxLength": 200,
            "title": "Department",
            "type": "string"
          },
          "durationSeconds": {
            "anyOf": [
              {
                "maximum": 50.0,
                "minimum": 3.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationseconds"
          },
          "language": {
            "default": "",
            "maxLength": 200,
            "title": "Language",
            "type": "string"
          },
          "platform": {
            "default": "",
            "maxLength": 200,
            "title": "Platform",
            "type": "string"
          },
          "purpose": {
            "default": "",
            "maxLength": 200,
            "title": "Purpose",
            "type": "string"
          },
          "ratio": {
            "anyOf": [
              {
                "enum": [
                  "1:1",
                  "3:4",
                  "4:3",
                  "9:16",
                  "16:9",
                  "21:9"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ratio"
          },
          "referenceImages": {
            "items": {
              "type": "string"
            },
            "maxItems": 16,
            "title": "Referenceimages",
            "type": "array"
          },
          "referenceVideos": {
            "items": {
              "type": "string"
            },
            "maxItems": 16,
            "title": "Referencevideos",
            "type": "array"
          },
          "requester": {
            "default": "",
            "maxLength": 200,
            "title": "Requester",
            "type": "string"
          },
          "requirementDescription": {
            "default": "",
            "maxLength": 4000,
            "title": "Requirementdescription",
            "type": "string"
          },
          "scene": {
            "default": "",
            "maxLength": 200,
            "title": "Scene",
            "type": "string"
          },
          "selling": {
            "default": "",
            "maxLength": 200,
            "title": "Selling",
            "type": "string"
          },
          "styleNos": {
            "items": {
              "maxLength": 64,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 20,
            "title": "Stylenos",
            "type": "array"
          },
          "theme": {
            "default": "",
            "maxLength": 200,
            "title": "Theme",
            "type": "string"
          },
          "videoType": {
            "default": "",
            "maxLength": 200,
            "title": "Videotype",
            "type": "string"
          }
        },
        "title": "TaskBrief",
        "type": "object"
      },
      "TaskCreateIn": {
        "additionalProperties": false,
        "description": "建一张需求单。比 ``TaskIn`` 多一个主款号。\n\n主款号只在创建时收：快照冻结之后就不许改写了，所以 ``PUT`` 用的还是 ``TaskIn``，\n往里塞 ``styleNo`` 会被 ``extra=\"forbid\"`` 挡成 422——想换款就提一张新的。",
        "properties": {
          "brief": {
            "$ref": "#/components/schemas/TaskBrief",
            "default": {
              "audience": "",
              "color": "",
              "contentType": "",
              "department": "",
              "language": "",
              "platform": "",
              "purpose": "",
              "referenceImages": [],
              "referenceVideos": [],
              "requester": "",
              "requirementDescription": "",
              "scene": "",
              "selling": "",
              "styleNos": [],
              "theme": "",
              "videoType": ""
            }
          },
          "deadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "priority": {
            "default": 0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Priority",
            "type": "integer"
          },
          "styleNo": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Styleno",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "styleNo"
        ],
        "title": "TaskCreateIn",
        "type": "object"
      },
      "TaskEnvelope": {
        "additionalProperties": false,
        "properties": {
          "task": {
            "$ref": "#/components/schemas/TaskOut"
          }
        },
        "required": [
          "task"
        ],
        "title": "TaskEnvelope",
        "type": "object"
      },
      "TaskIn": {
        "additionalProperties": false,
        "description": "建一张需求单，或整体覆盖一张已有的。\n\n创建与更新用同一个形状：更新就是整体覆盖（PUT），不做字段级的局部合并——局部合并\n在「哪些字段发布后冻结」这种规则面前很难说清楚「没传」到底是不改还是清空。",
        "properties": {
          "brief": {
            "$ref": "#/components/schemas/TaskBrief",
            "default": {
              "audience": "",
              "color": "",
              "contentType": "",
              "department": "",
              "language": "",
              "platform": "",
              "purpose": "",
              "referenceImages": [],
              "referenceVideos": [],
              "requester": "",
              "requirementDescription": "",
              "scene": "",
              "selling": "",
              "styleNos": [],
              "theme": "",
              "videoType": ""
            }
          },
          "deadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "priority": {
            "default": 0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Priority",
            "type": "integer"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "TaskIn",
        "type": "object"
      },
      "TaskOut": {
        "additionalProperties": false,
        "properties": {
          "assigneeUserIds": {
            "default": [],
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Assigneeuserids",
            "type": "array"
          },
          "brief": {
            "$ref": "#/components/schemas/TaskBrief"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "creatorUserId": {
            "format": "uuid",
            "title": "Creatoruserid",
            "type": "string"
          },
          "deadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "style": {
            "$ref": "#/components/schemas/TaskStyle"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "priority",
          "deadline",
          "creatorUserId",
          "style",
          "brief",
          "createdAt",
          "updatedAt"
        ],
        "title": "TaskOut",
        "type": "object"
      },
      "TaskStyle": {
        "additionalProperties": false,
        "description": "下单那天主款长什么样，创建时冻结。\n\n抄一份而不是每次回头查产品资料：上游随时改名换图，历史需求单不该跟着变样。除了\n款号，另三项在上游缺名缺图时是空字符串。",
        "properties": {
          "brand": {
            "default": "",
            "maxLength": 200,
            "title": "Brand",
            "type": "string"
          },
          "category": {
            "default": "",
            "maxLength": 200,
            "title": "Category",
            "type": "string"
          },
          "previewImageUrl": {
            "default": "",
            "title": "Previewimageurl",
            "type": "string"
          },
          "styleNo": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Styleno",
            "type": "string"
          }
        },
        "required": [
          "styleNo"
        ],
        "title": "TaskStyle",
        "type": "object"
      },
      "TasksPageOut": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "TasksPageOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "iclip_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "iclip-agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/conversations/{conversation_id}/task": {
      "put": {
        "operationId": "set_conversation_task_conversations__conversation_id__task_put",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationTaskIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Conversation Task",
        "tags": [
          "conversations"
        ]
      }
    },
...
    "/tasks": {
      "get": {
        "operationId": "list_tasks_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "draft",
                    "published",
                    "confirmed",
                    "withdrawn"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "claimedBy",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^me$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Claimedby"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TasksPageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tasks",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "operationId": "create_task_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{task_id}": {
      "delete": {
        "operationId": "delete_task_tasks__task_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Task",
        "tags": [
          "tasks"
        ]
      },
      "get": {
        "operationId": "get_task_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task",
        "tags": [
          "tasks"
        ]
      },
      "put": {
        "operationId": "save_task_tasks__task_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{task_id}/confirm": {
      "post": {
        "operationId": "confirm_task_tasks__task_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{task_id}/publish": {
      "post": {
        "operationId": "publish_task_tasks__task_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{task_id}/withdraw": {
      "post": {
        "operationId": "withdraw_task_tasks__task_id__withdraw_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Withdraw Task",
        "tags": [
          "tasks"
        ]
      }
    },
...
  }
}